+++ Server/CreateEntityMessage.h
@@ -18,7 +18,7 @@ struct CreateEntityMessage : public Message
     float posY;
     float posZ;
 
-    int getId() const override;
+    byte getId() const override;
 
     std::vector<uint8_t>& serialize(Serializer& serializer) const override;
     void deserialize(Deserializer& deserializer) override;
