+++ Server/Message.h
@@ -11,7 +11,7 @@ struct Message
 {
     Message() = default;
 
-    virtual int getId() const = 0;
+    virtual byte getId() const = 0;
     virtual std::vector<uint8_t>& serialize(Serializer& serializer) const = 0;
     virtual void deserialize(Deserializer& deserializer) = 0;
     virtual void process(const sockaddr_in& senderAddr)  = 0;
