+++ Server/MoveEntityMessage.h
@@ -16,7 +16,7 @@ struct MoveEntityMessage : public Message
     float inputX;
     float inputY;
 
-    int getId() const override;
+    byte getId() const override;
 
     std::vector<uint8_t>& serialize(Serializer& serializer) const override;
     void deserialize(Deserializer& deserializer) override;
