+++ Server/RotateEntityMessage.h
@@ -16,7 +16,7 @@ struct RotateEntityMessage : public Message
     float mouseInputX;
     float mouseInputY;
 
-    int getId() const override;
+    byte getId() const override;
 
     std::vector<uint8_t>& serialize(Serializer& serializer) const override;
     void deserialize(Deserializer& deserializer) override;
