+++ Server/SynchronizeEntitiesMessage.h
@@ -9,7 +9,7 @@ struct SynchronizeEntitiesMessage : public Message
 
     SynchronizeEntitiesMessage();
 
-    int getId() const override;
+    byte getId() const override;
     std::vector<uint8_t>& serialize(Serializer& serializer) const override;
     void deserialize(Deserializer& deserializer) override;
     void process(const sockaddr_in& senderAddr) override;
